Changes:
* Added Tolk library in replacement of ScreenReaderAPI.
* Solved some problems related to resource loading.

+++ soundrts/world.py
@@ -336,7 +336,7 @@ class World(object):
                 square.high_ground = square.name in self.high_grounds
         self.set_neighbours()
         xmax = self.nb_columns * self.square_width
-        res = COLLISION_RADIUS * 2 / 3
+        res = int(COLLISION_RADIUS * 2 / 3)
         self.collision = {"ground": collision.CollisionMatrix(xmax, res),
                           "air": collision.CollisionMatrix(xmax, res)}
